// src/modules/imports/types/SpreadsheetType.ts
// src/modules/imports/types/SpreadsheetType.ts
export enum SpreadsheetType {
  CHECKLIST_INDUSTRIA = 'CHECKLIST_INDUSTRIA',
  ROTEIRO_PROMOTORES = 'ROTEIRO_PROMOTORES',
  FREQUENCIA_LOJAS = 'FREQUENCIA_LOJAS',
  PROMOTORES = 'PROMOTORES',
  LOJAS = 'LOJAS',
  INDUSTRIAS = 'INDUSTRIAS',
  DESCONHECIDO = 'DESCONHECIDO'
}